+++ R/outsider.R
@@ -24,7 +24,7 @@ NULL
 #' @param show_docker Show docker messages? Default FALSE.
 #' @details For more control see \code{\link[outsider.base]{log_set}}
 #' @return data.frame
-#' @examples examaples/module_install.R
+#' @example examples/module_install.R
 #' @export
 verbosity_set <- function(show_program = TRUE,
                           show_docker = FALSE) {
@@ -33,6 +33,33 @@ verbosity_set <- function(show_program = TRUE,
   log_set(log = 'docker_out', val = show_docker)
   log_set(log = 'docker_err', val = show_docker)
   invisible(TRUE)
+}
+
+#' @name ssh_setup
+#' @title Setup SSH
+#' @description Send all outsider commands to an external host. Provide an
+#' \code{ssh} session to this function and all subsequent commands will be run
+#' on the host rather than the local machine. When finished it is always good
+#' practice to disconnect from the remote host by running \code{ssh_teardown}.
+#' @param session \code{ssh} session, see \code{\link[ssh]{ssh_connect}}
+#' @return logical
+#' @example examples/ssh.R
+#' @export
+ssh_setup <- function(session) {
+  res <- server_connect(session)
+  message(paste0('Remember to run ', func('ssh_teardown'), ' when finished.'))
+  res
+}
+
+#' @name ssh_teardown
+#' @title Teardown SSH
+#' @description Disconnect from a remote host and stop commands being
+#' transferred.
+#' @return logical
+#' @example examples/ssh.R
+#' @export
+ssh_teardown <- function() {
+  server_disconnect()
 }
 
 .onAttach <- function(...) {