+++ examples/ssh.R
@@ -1,0 +1,7 @@
+library(outsider)
+
+# # NOT RUN
+# session <- ssh::ssh_connect(host = "[IP ADDRESS]")
+# ssh_setup(session)
+# # (Run outsider commands)
+# ssh_teardown()